requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
passlib[bcrypt]
python-jose[cryptography]
jinja2
python-multipart
asyncpg
python-dotenv
pydantic-settings
